Started work on rtl provider
