+++ client/packages/components/src/index.tsx
@@ -1,19 +1,14 @@
 import React from 'react'
-import { createMuiTheme, CssBaseline, ThemeProvider } from '@material-ui/core'
+import { CssBaseline } from '@material-ui/core'
 import { HashRouter as Router, Switch, Route, Redirect } from 'react-router-dom'
 import { CustomAppBar } from './layout/AppBar'
 import { Container } from './layout/Container'
 import { Login, Register, Classes } from './containers'
 import { WithTranslateProps } from './external-types'
-
-const theme = createMuiTheme({
-  typography: {
-    fontFamily: '"Heebo", "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif'
-  }
-})
+import { RtlProvider } from './providers'
 
 export const Application = (i18nProps: WithTranslateProps) => (
-  <ThemeProvider theme={theme}>
+  <RtlProvider>
     <CssBaseline />
     <Router>
       <CustomAppBar {...i18nProps} />
@@ -38,5 +33,5 @@ export const Application = (i18nProps: WithTranslateProps) => (
         </Switch>
       </Container>
     </Router>
-  </ThemeProvider>
+  </RtlProvider>
 )
